no more mandatory dependency on bdd_vars (but integrates with play_selenium, if available)
