+++ tests/test_plugin.py
@@ -18,17 +18,7 @@ def test_play_engine_class(play_engine_class):
     assert play_engine_class is PlayEngine
 
 
-def test_play(play, navigation, bdd_vars):
+def test_play(play, navigation):
     assert play.navigation is navigation
     assert play.navigation.page is navigation.page
-    assert play.variables != bdd_vars
-    assert 'base_url' in play.variables
-    assert 'base_url' not in bdd_vars
-
-
-def test_play_variables(play, navigation, bdd_vars):
-    """ If you provide values inside a pytest-play section of your pytest-variables
-        file, they become available to pytest-play """
-    assert 'date_format' in play.variables
-    assert 'date_format' not in bdd_vars
-    assert play.variables['date_format'] == 'YYYYMMDD'
+    assert play.variables['base_url'] == 'http://'
