+++ tests/test_plugin_bdd_vars.py
@@ -1,0 +1,28 @@
+# -*- coding: utf-8 -*-
+import pytest
+
+
+@pytest.fixture
+def browser():
+    import mock
+    browser = mock.MagicMock()
+
+    from zope.interface import alsoProvides
+    from pypom.splinter_driver import ISplinter
+    alsoProvides(browser, ISplinter)
+    return browser
+
+
+@pytest.fixture
+def bdd_vars():
+    """ Simulate integration with play_selenium """
+    return {'using': 'bdd_vars'}
+
+
+def test_play_variables(play, navigation, bdd_vars):
+    """ If you provide values inside a pytest-play section of your pytest-variables
+        file, they become available to pytest-play """
+    assert 'date_format' in play.variables
+    assert 'date_format' not in bdd_vars
+    assert play.variables['date_format'] == 'YYYYMMDD'
+    assert play.variables['using'] == 'bdd_vars'
